# backend/docs/swagger.yaml
basePath: /
definitions:
  models.Bike:
    properties:
      bike_id:
        type: string
      latitude:
        type: number
      longtitude:
        type: number
      name:
        type: string
      rented:
        type: boolean
      user_id:
        type: string
    type: object
  models.BikeResponseStruct:
    properties:
      bike_id:
        type: string
      latitude:
        type: number
      longtitude:
        type: number
      name:
        type: string
      rented:
        type: boolean
    type: object
  models.LoginRequest:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
  models.Response:
    properties:
      code:
        example: 418
        type: integer
      message:
        example: A simple message briefly explaining the issue
        type: string
    type: object
  models.User:
    properties:
      bike_id:
        type: string
      name:
        type: string
      renting:
        type: boolean
      user_id:
        type: string
    type: object
  models.UserResponseStruct:
    properties:
      bike_id:
        type: string
      name:
        type: string
      renting:
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  title: Bike renting service API
  version: 1.0.0
paths:
  /bikes:
    get:
      description: Fetches all bikes from the database.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Bike'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Fetches all bikes from the database.
      tags:
      - Bikes
  /bikes/:id:
    get:
      consumes:
      - text/plain
      description: Fetches a bike with a user-provided ID.
      parameters:
      - description: bike_id in models.Bike
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BikeResponseStruct'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Fetches a bike with a user-provided ID.
      tags:
      - Bikes
  /fetchUserData:
    get:
      description: Fetches the data of the logged in user.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Fetches the data of the logged in user.
      tags:
      - Users
  /login:
    post:
      consumes:
      - application/json
      description: Handles the login procedure to the service.
      parameters:
      - description: name and password values in models.User
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserResponseStruct'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Handles the login procedure to the service.
      tags:
      - Authorization
  /logout:
    get:
      description: Allows the user to log out from the service.
      responses:
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Allows the user to log out from the service.
      tags:
      - Authorization
  /rent/:id:
    patch:
      consumes:
      - text/plain
      description: User rents the bike.
      parameters:
      - description: bike_id in models.Bike
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: User rents the bike.
      tags:
      - Bikes
  /return/:id:
    patch:
      consumes:
      - text/plain
      description: User returns the bike.
      parameters:
      - description: bike_id in models.Bike
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: User returns the bike.
      tags:
      - Bikes
swagger: "2.0"